// WeddingService_AdminDashboard/src/app/modules/user/modules/super-admin/modules/vendor-management/model/gst-detail.model.ts
export class GstDetailModel {
    id: number;
    kycid: number;
    gstnumber: string;
    gststate: string;
    gststateName: string;
    gstcity: string;
    gstcityName: string;
    gstdocument: string;
    businessName: string;
    address: string;
    active: number;
    createdBy: number;
    createdAt: string;
    updatedBy: number;
    updatedAt: null
}